Create watcher chan per applog
Somehow I was lucky that my previous fixed seemed to work: https://github.com/symfony-cli/symfony-cli/pull/1 .

After testing the release version it never worked.

Turns out, for every `--file` there is a goroutine that reads from `watcherChan`.

The problem is, that when a goroutine receives a message it cannot handle, it discards it:
https://github.com/symfony-cli/symfony-cli/blob/cc0df3b6110d4c5657a158c8363a301e6703fa44/local/logs/tailer.go#L144-L151

Therefore, the other goroutines never receive their message.

To solve this, we create a channel per log. I think this is what was the intention.