Convert relative paths to absolute paths when tailing
This makes it possible to run `symfony server:log --file=var/dev.log` while
the `var/dev.log` does not exist yet. When the file is created after the tailer
has been started, inotify will emit an event where `e.Path()` is an absolute path.

We need to compare that with the absolute path to be able to start the tailer the first time.