Evaluate symlinks before starting tailer
When the current directory or the absolute path provided to `--file` is a symlink, we need
to convert it to the realpath on the disk.

The reason for this, is that inotify will always emit events with the realpath and not the symlink
path.

Without this change, when the log files exist on the time of starting the tail, everything will be fine.

But when you define log files that do not yet exist at the time of starting, they will never be picked up
if they are located in a symlink path.