Fix the PATH value when executing PHP, fix #124
Previously, we were prepending the full path to the PHP binary instead of the directory containing it. We now prepend the directory instead.
Additionally we now preprend to PATH the special temporary directory created with symlinks to PHP binaries which should allow to make it work if PHP binary is not named 'php'.