Add support for PHP streaming support
Relatively easy to implement for PHP builtin but required to change the way we handle CGI and FPM.
We now use the golang reverse proxy also for these use cases but tweaking the RoundTripper to communicate over CGI instead.
This means we also had to change the way we inspect the responses for the WDT of X-Sendfile because now the original Body needs to be closed in a proper way and at the right time.

-