Rework workers lifecycle management to make them stop cleanly
The workers will already receive the SIGINT signal because Ctrl+C is sent to the whole process group in which they are.
This means that in the `server:start` command we only need to disable worker restart and wait for them to finish before exiting.
In the `server:stop` command, this means we have to first signal the monitoring process (but without stopping it) and then stop each worker individually.

Fix #390