feat: allow to override the path to the Symfony Console
When running `symfony console` we are currently looking for a `bin/console` or a `app/console` file to run.
What if the user renamed the file? or if they have a non "stantard" project structure?

This PR intends to introduce a way one can override the path to the console by defining `SYMFONY_CONSOLE_PATH` environment variable.