do not truncate log files on each LogWriter calls
We used to have a single concurrent usage of the `LogWriter`, but this is not the case anymore since b29665f7afcba0a71a30077d3516aa6076908924.
Because of this we now loose the runner logs as he file is being truncated during the `buildCmd` call when the LogWrited is fetched.

To fix this i propose to instantiate the LogWriter once which will effectively truncate the file only once per `server:start` launch.