envs: do not try to parse the query into something valid.
Symfony Cloud systems use the "query" for other things than the query
parameters used for the URL, so we cannot rely on it. This was trying
to parse it into a `map[string]string` which is just
incompatible.

This PR drops the part that was trying to make it too smart and keeps
it simpler.